Move redactor from bootstrap to new redaction module
