+++ redaction/redactor.go
@@ -1,9 +1,16 @@
-package bootstrap
+package redaction
 
 import (
 	"bytes"
 	"io"
 )
+
+// RedactLengthMin is the shortest string length that will be considered a
+// potential secret by the environment redactor. e.g. if the redactor is
+// configured to filter out environment variables matching *_TOKEN, and
+// API_TOKEN is set to "none", this minimum length will prevent the word "none"
+// from being redacted from useful log output.
+const RedactLengthMin = 6
 
 type Redactor struct {
 	replacement []byte
